Completed Simple Color Flipper
Created first portion: simple color flipper in index.html and added functionality in app.js.
Cycles through a set of 4 fixed color values, displays the color value in text and sets the background to the color.
Has two main functions: One that chooses the color at random and another that cycles through the colors in order.